requirements.txt
cbor
requests
beautifulsoup4
lxml
scrapy
psycopg2
redis
pyyaml